Add optional tags display to Posts component
Enabled conditional rendering of tags in the Posts component through a new showTags prop. This allows tag visibility to be toggled depending on where the Posts component is used, enhancing customization and meeting design requirements where tag display is not always desired. The PostsPage now takes advantage of this flexibility by opting to show tags.