Add 'About' page and navigation links
Introduced a new 'About' page to the website providing personal narrative and development insights of the creator, Jim Horn. Navigation links to this page have been added to both the header menu and the footer for better discoverability. This enriches the site with a more personal touch and offers users an overview of the behind-the-scenes and technical choices involved in building the site.